+++ ui/web_frontend/src/components/viewer/DOMUtils.js
