+++ ui/web_frontend/src/components/viewer/ViewNote.vue
@@ -16,7 +16,7 @@ import { useI18n } from 'vue-i18n'
 import { MdEditor } from 'md-editor-v3'
 import { ElMessage } from 'element-plus'
 import AddDialog from '@/components/datatable/AddDialog.vue'
-import { getDefaultPath, getDefaultVault } from './dataUtils';
+import { getDefaultPath, getDefaultVault } from '@/components/datatable/dataUtils';
 
 const emit = defineEmits(['noteChange'])
